# OpenAPI接口.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/admin/list": {
      "get": {
        "summary": "查询用户列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理端接口"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "用户ID"
                          },
                          "email": {
                            "type": "string",
                            "description": "用户邮箱"
                          },
                          "role": {
                            "type": "string",
                            "description": "用户角色（管理员/用户）"
                          },
                          "nickname": {
                            "type": "string",
                            "description": "用户昵称"
                          },
                          "quotaSpace": {
                            "type": "integer",
                            "description": "空间配额 单位：字节"
                          },
                          "quotaCount": {
                            "type": "integer",
                            "description": "可上传文件数量 单位：个"
                          },
                          "usedSpace": {
                            "type": "integer",
                            "description": "已使用空间"
                          },
                          "usedCount": {
                            "type": "integer",
                            "description": "已上传数量"
                          },
                          "createTime": {
                            "type": "string"
                          },
                          "updateTime": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 1,
                  "message": "success",
                  "data": [
                    {
                      "id": 6,
                      "email": "xingyuan_guo@outlook.com",
                      "role": "ADMIN",
                      "nickname": "管理员",
                      "quotaSpace": 104857600,
                      "quotaCount": 100,
                      "usedSpace": 0,
                      "usedCount": 0,
                      "createTime": "2025-12-01 18:09:22",
                      "updateTime": "2025-12-01 20:53:44"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/disableUser": {
      "post": {
        "summary": "禁用用户账号",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理端接口"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "durationDays": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userId",
                  "reason",
                  "durationDays"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 1,
                  "message": "success",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/enableUser/{id}": {
      "put": {
        "summary": "启用用户",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理端接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 7,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 1,
                  "message": "success",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/deleteUser/{id}": {
      "delete": {
        "summary": "删除用户",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理端接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 7,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 1,
                  "message": "success",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/updateSpace": {
      "post": {
        "summary": "修改用户空间配额",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理端接口"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "quotaSpace": {
                    "type": "integer"
                  },
                  "quotaCount": {
                    "type": "integer"
                  }
                },
                "required": [
                  "id",
                  "quotaSpace",
                  "quotaCount"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 1,
                  "message": "success",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/searchUser": {
      "post": {
        "summary": "搜索用户",
        "deprecated": false,
        "description": "根据条件筛选用户，返回用户列表",
        "tags": [
          "管理端接口"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "邮箱"
                  },
                  "nickname": {
                    "type": "string",
                    "description": "昵称"
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "账号状态"
                  }
                }
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "用户ID"
                          },
                          "email": {
                            "type": "string",
                            "description": "邮箱"
                          },
                          "role": {
                            "type": "string",
                            "description": "角色"
                          },
                          "nickname": {
                            "type": "string",
                            "description": "昵称"
                          },
                          "enabled": {
                            "type": "boolean",
                            "description": "账号状态"
                          },
                          "quotaSpace": {
                            "type": "integer",
                            "description": "总空间配额"
                          },
                          "quotaCount": {
                            "type": "integer",
                            "description": "总可上传次数"
                          },
                          "usedSpace": {
                            "type": "integer",
                            "description": "已使用空间"
                          },
                          "usedCount": {
                            "type": "integer",
                            "description": "已上传次数"
                          },
                          "disabledUntil": {
                            "type": "string",
                            "description": "解封时间"
                          },
                          "disableReason": {
                            "type": "string",
                            "description": "被封禁原因"
                          },
                          "createTime": {
                            "type": "string",
                            "description": "创建时间"
                          },
                          "updateTime": {
                            "type": "string",
                            "description": "更新时间"
                          }
                        },
                        "required": [
                          "id",
                          "email",
                          "role",
                          "nickname",
                          "enabled",
                          "quotaSpace",
                          "quotaCount",
                          "usedSpace",
                          "usedCount",
                          "disabledUntil",
                          "disableReason",
                          "createTime",
                          "updateTime"
                        ]
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/getUser/{id}": {
      "get": {
        "summary": "根据id搜索用户",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理端接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 8,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "enabled": {
                          "type": "boolean"
                        },
                        "quotaSpace": {
                          "type": "integer"
                        },
                        "quotaCount": {
                          "type": "integer"
                        },
                        "usedSpace": {
                          "type": "integer"
                        },
                        "usedCount": {
                          "type": "integer"
                        },
                        "disabledUntil": {
                          "type": "null"
                        },
                        "disableReason": {
                          "type": "null"
                        },
                        "createTime": {
                          "type": "string"
                        },
                        "updateTime": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "role",
                        "nickname",
                        "enabled",
                        "quotaSpace",
                        "quotaCount",
                        "usedSpace",
                        "usedCount",
                        "disabledUntil",
                        "disableReason",
                        "createTime",
                        "updateTime"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 1,
                  "message": "success",
                  "data": {
                    "id": 8,
                    "email": "2569856768@qq.com",
                    "role": "USER",
                    "nickname": "构型元",
                    "enabled": true,
                    "quotaSpace": 104857600,
                    "quotaCount": 100,
                    "usedSpace": 0,
                    "usedCount": 0,
                    "disabledUntil": null,
                    "disableReason": null,
                    "createTime": "2025-12-01 21:10:10",
                    "updateTime": "2025-12-01 21:13:42"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/updateTestStatus": {
      "put": {
        "summary": "修改测试状态",
        "deprecated": false,
        "description": "",
        "tags": [
          "管理端接口"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "testStatus": {
                    "type": "boolean",
                    "description": "测试状态"
                  },
                  "testMessage": {
                    "type": "string",
                    "description": "测试信息"
                  }
                },
                "required": [
                  "testStatus",
                  "testMessage"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}